requirements.txt
Django==1.9.12
Faker==0.7.16
Markdown==2.6.8
Pillow==4.0.0
PyJWT==1.4.2
appdirs==1.4.2
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
factory-boy==2.8.1
markdown2==2.3.3
olefile==0.44
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
